[GraduationTimes] Move to components/material for reuse in Study programme
