+++ services/frontend/src/components/material/FacultyGraduations.tsx
