+++ services/frontend/src/components/material/GraduationTimes/MedianDisplay/MedianBarChart.tsx
